+++ tests/e2e/a11y/home.a11y.spec.ts
@@ -1,6 +1,6 @@
 import { test } from '@tests/fixtures'
 import { languageKeys } from '@i18n/ui'
-import { userInHome } from '@tests/pages/A11yUserJourney'
+import { userInHome } from '@tests/pages/a11y/A11yUserJourney'
 
 
 test.describe('A11y - Home', { tag: ['@a11y', '@home'] }, () => {
