+++ tests/e2e/flow/charla.flow.spec.ts
@@ -1,6 +1,6 @@
 import { test } from '@tests/fixtures'
 import { languageKeys, type UILanguages } from '@i18n/ui'
-import { userInTalkList } from '@tests/pages/TalkUserJourney'
+import { userInTalkList } from '@tests/pages/content/TalkUserJourney'
 
 const SLUG = '2023-09-27-devcontainers'
