+++ tests/e2e/functional/homepage.language.spec.ts
@@ -1,6 +1,6 @@
 import { test } from '@tests/fixtures'
 import type { UILanguages } from '@i18n/ui'
-import { userInHome } from '@tests/pages/HomeLanguageUserJourney'
+import { userInHome } from '@tests/pages/home/HomeLanguageUserJourney'
 
 type LocaleSwitchFixture = {
   from: UILanguages
