+++ tests/e2e/functional/theme/theme-persistence.spec.ts
@@ -1,5 +1,5 @@
 import { test } from '@tests/fixtures'
-import { userInHome } from '@tests/pages/ColorSwitchUserJourney'
+import { userInHome } from '@tests/pages/home/ColorSwitchUserJourney'
 
 test.describe('Theme local storage',
   { tag: ['@theme-local-storage', '@functional'] },
