+++ tests/pages/shared/NavigationPaths.ts
