+++ tests/pages/sidebar/ThemeToggle.ts
@@ -1,6 +1,6 @@
 import { step } from '@tests/fixtures'
-import { target } from '@tests/pages/Target'
-import type { Target as TargetComponent } from '@tests/pages/Target'
+import { target } from '@tests/pages/components/Target'
+import type { Target as TargetComponent } from '@tests/pages/components/Target'
 import type { Page } from '@playwright/test'
 
 export class ThemeToggle {
